restructure proejct with multiple build files
