--- BUILD
@@ -1,6 +1,0 @@
-load("@rules_java//java:defs.bzl", "java_binary")
-
-java_binary(
-    name = "DECIDE",
-    srcs = glob(["src/main/java/com/**/*.java"]),
-)
