+++ src/main/java/com/example/BUILD
@@ -1,0 +1,3 @@
+package(
+    default_visibility = ["//src:__subpackages__"],
+)
