--- src/main/java/com/example/DECIDE.java
@@ -1,7 +1,0 @@
-package com.example;
-
-public class DECIDE {
-    public static void main(String args[]) {
-        main.java.com.example.Greeting.sayHi();
-    }
-}
