+++ src/main/java/com/example/cmdline/BUILD
@@ -1,0 +1,8 @@
+load("@rules_java//java:defs.bzl", "java_binary")
+
+java_binary(
+    name = "DECIDE",
+    srcs = ["DECIDE.java"],
+    main_class = "com.example.cmdline.DECIDE",
+    deps = ["//src/main/java/com/example/greeter:greeter"],
+)
