+++ src/main/java/com/example/cmdline/DECIDE.java
@@ -1,0 +1,9 @@
+package com.example.cmdline;
+
+import com.example.greeter.Greeting;
+
+public class DECIDE {
+    public static void main(String args[]) {
+        Greeting.sayHi();
+    }
+}
