+++ src/main/java/com/example/greeter/BUILD
@@ -1,0 +1,10 @@
+package(
+    default_visibility = ["//src:__subpackages__"],
+)
+
+java_library(
+    name = "greeter",
+    srcs = glob([
+       "*.java",
+   ]),
+)
