--- test/java/GreetingTest.java
@@ -1,20 +1,0 @@
-import org.junit.Test;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.rules.Timeout;
-import static org.junit.Assert.*;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.CoreMatchers.*;
-
-public class GreetingTest {
-    @Before
-    public void setUp() {
-
-    }
-
-    @Test
-    public void insertIsTrueWhenElementIsNotInTree() {
-        assertTrue(true);
-    }
-}
